requirements.txt
apache-airflow-providers-github
apache-airflow-providers-snowflake